refactor: improve directory structure to keep components into related modules
