+++ app/page.tsx
@@ -3,7 +3,7 @@ import Footer from "../components/gernal/footer";
 import HeroSection from "../components/landing-page/hero-section";
 import HowWeWork from "../components/landing-page/how-we-work";
 import OurPartners from "../components/landing-page/our-partners";
-import Portfolio from "../components/landing-page/portfolio";
+import Portfolio from "../components/portfolio/portfolio";
 import Technologies from "../components/landing-page/technologies";
 import Testimonials from "../components/landing-page/testimonials";
 import WhoWeAre from "../components/landing-page/who-we-are";
